+++ src/pages/mission/droner/MissionReport.tsx
@@ -23,11 +23,24 @@ import { CampaignDatasource } from "../../../datasource/CampaignDatasource";
 import { DateTimeUtil } from "../../../utilities/DateTimeUtil";
 import { numberWithCommas } from "../../../utilities/TextFormatter";
 import { RedeemDronerListEntity } from "../../../entities/RedeemEntities";
+import styled from "styled-components";
 const _ = require("lodash");
+
+const NewTable = styled(Table)<{
+  colors: string;
+}>`
+  .ant-table-container table thead tr th {
+    background-color: ${({ colors }) => colors} !important;
+    font-family: "Prompt" !important;
+    font-weight: 500 !important;
+    color: white !important;
+    font-weight: bold !important;
+  }
+`;
 
 function MissionReport() {
   let queryString = _.split(window.location.pathname, "=");
-  const [type, setType] = useState("unsuccess");
+  const [type, setType] = useState("unsuccessMission");
   const row = 10;
   const rowCard = 5;
   const [num, setNum] = useState<number>(1);
@@ -44,16 +57,21 @@ function MissionReport() {
   const [dataCondition, setDataCondition] = useState<ConditionMission[]>();
   const [startDate, setStartDate] = useState<any>(null);
   const [endDate, setEndDate] = useState<any>(null);
+  const [statusMission, setStatusMission] = useState("INPROGRESS");
+  const [isLoading, setIsLoading] = useState(false);
 
   const fetchMissionInprogress = () => {
+    setIsLoading(true);
     CampaignDatasource.detailMissionInprogress(
       queryString[1],
       num,
       rowCard,
-      current
+      current,
+      statusMission
     ).then((res) => {
       setDataCondition(res.condition);
       setDataMission(res);
+      setIsLoading(false);
     });
   };
 
@@ -65,8 +83,8 @@ function MissionReport() {
       current,
       dataMission?.missionNo
     ).then((res) => {
-      console.log(res);
       setDataMissionSuc(res);
+      setIsLoading(false);
     });
   };
 
@@ -77,7 +95,8 @@ function MissionReport() {
   useEffect(() => {
     fetchMissionInprogress();
     fetchMissionSuccess();
-  }, [current, num]);
+    
+  }, [current, num, statusMission]);
 
   const handleSearchDate = (e: any) => {
     if (e != null) {
@@ -116,6 +135,18 @@ function MissionReport() {
       return 0;
     }
     return a > b ? 1 : -1;
+  };
+
+  const mapColor: any = {
+    unsuccessMission: "rgba(235, 87, 87, 0.1)",
+    unconfirmMission: "rgba(255, 250, 235, 1)",
+    successMission: "rgba(33, 150, 83, 0.1)",
+  };
+
+  const mapTableColor: any = {
+    unsuccessMission: "#EB5757",
+    unconfirmMission: "#FFCA37",
+    successMission: color.Success,
   };
 
   const columns = [
@@ -158,7 +189,7 @@ function MissionReport() {
         };
       },
     },
-    type === "unsuccess"
+    type === "unsuccessMission"
       ? {
           title: "จำนวนไร่สะสม",
           dataIndex: "allraiAmount",
@@ -166,12 +197,12 @@ function MissionReport() {
           sorter: (a: any, b: any) => sorter(a.allraiAmount, b.allraiAmount),
           render: (value: any, row: any, index: number) => {
             const calRai = () => {
-              return numberWithCommas(
+              let cal = (
                 dataCondition?.find((x) => x.num === num)?.rai! -
-                  row.allraiAmount
-              );
+                row.allraiAmount
+              ).toFixed(2);
+              return numberWithCommas(Number(cal));
             };
-
             return {
               children: (
                 <div>
@@ -227,29 +258,40 @@ function MissionReport() {
         }}
       >
         <h5 style={{ color: "white", alignSelf: "center" }}>
-          ผู้เข้าร่วมภารกิจ : {dataMission?.amoutPeople} คน
+          ผู้เข้าร่วมภารกิจ : {dataMission?.amountPeople} คน
         </h5>
       </div>
-      {dataMission?.condition.map((item: any, index: any) => (
-        <div
-          className="pt-3"
-          key={index}
-          onClick={() => {
-            setNum(item.num);
-          }}
-          style={{ cursor: "pointer" }}
-        >
-          <MissionReportCard
-            checkCard={checkSubmission(index + 1)}
-            title={`ภารกิจ ${item.num}  ${item.missionName}`}
-            raiAmount={item.rai}
-            successPoint={item.reward.amountSuccessCount}
-            unsuccessPoint={item.reward.amountInprogressCount}
-            img={item.reward.imagePath}
-            missionName={`${item.reward.rewardName} | ${item.reward.rewardNo} | ${item.reward.rewardType}  (${item.reward.rewardExchange}) | คงเหลือ ${item.reward.remain} ชิ้น`}
-          />
-        </div>
-      ))}
+      {dataMission?.condition.map((item: any, index: any) => {
+        const detailReward: any = {
+          rewardId: item.reward.id,
+          rewardName: item.reward.rewardName,
+          rewardNo: item.reward.rewardNo,
+          rewardExchange: item.reward.rewardExchange,
+          remain: item.reward.remain,
+          imagePath: item.reward.imagePath,
+          rewardType: item.reward.rewardType,
+        };
+        return (
+          <div
+            className="pt-3"
+            key={index}
+            onClick={() => {
+              setNum(item.num);
+            }}
+            style={{ cursor: "pointer" }}
+          >
+            <MissionReportCard
+              checkCard={checkSubmission(index + 1)}
+              title={`ภารกิจ ${item.num}  ${item.missionName}`}
+              raiAmount={item.rai}
+              successMission={item.reward.amountSuccessCount}
+              unsuccessMission={item.reward.amountInprogressCount}
+              unconfirmMission={item.reward.amountRequestCount}
+              detailReward={detailReward}
+            />
+          </div>
+        );
+      })}
       <div className="d-flex justify-content-between pt-3 pb-3">
         <p>รายการทั้งหมด {dataMission?.condition.length} รายการ</p>
         <Pagination
@@ -279,17 +321,38 @@ function MissionReport() {
             borderBottomLeftRadius: 5,
             borderTopLeftRadius: 5,
             backgroundColor:
-              type === "unsuccess" ? "rgba(235, 87, 87, 0.1)" : color.White,
-            color: type === "unsuccess" ? color.Error : color.BK,
-            borderColor: type === "unsuccess" ? color.Error : color.BK,
-            borderWidth: type === "unsuccess" ? 1 : 0,
+              type === "unsuccessMission" ? mapColor[type] : color.White,
+            color: type === "unsuccessMission" ? color.Error : color.BK,
+            borderColor: type === "unsuccessMission" ? color.Error : color.BK,
+            borderWidth: type === "unsuccessMission" ? 1 : 0,
           }}
-          value="unsuccess"
+          value="unsuccessMission"
+          onClick={() => setStatusMission("INPROGRESS")}
         >
           ผู้เข้าร่วมที่ยังไม่สำเร็จ{" "}
           {`(${
             dataCondition?.find((x) => x.num === num)?.reward
               .amountInprogressCount
+          })`}
+        </Radio.Button>
+        <Radio.Button
+          className="col"
+          style={{
+            textAlign: "center",
+            padding: 4,
+            height: "40px",
+            backgroundColor:
+              type === "unconfirmMission" ? mapColor[type] : color.White,
+            color: type === "unconfirmMission" ? "#FFCA37" : color.BK,
+            borderColor: type === "unconfirmMission" ? "#FFCA37" : color.BK,
+            borderWidth: type === "unconfirmMission" ? 1 : 0,
+          }}
+          value="unconfirmMission"
+          onClick={() => setStatusMission("REQUEST")}
+        >
+          ยังไม่กดยืนยัน{" "}
+          {`(${
+            dataCondition?.find((x) => x.num === num)?.reward.amountRequestCount
           })`}
         </Radio.Button>
         <Radio.Button
@@ -301,12 +364,12 @@ function MissionReport() {
             textAlign: "center",
             padding: 4,
             backgroundColor:
-              type === "success" ? "rgba(33, 150, 83, 0.1)" : color.White,
-            color: type === "success" ? color.Success : color.BK,
-            borderColor: type === "success" ? color.Success : color.BK,
-            borderWidth: type === "success" ? 1 : 0,
+              type === "successMission" ? mapColor[type] : color.White,
+            color: type === "successMission" ? color.Success : color.BK,
+            borderColor: type === "successMission" ? color.Success : color.BK,
+            borderWidth: type === "successMission" ? 1 : 0,
           }}
-          value="success"
+          value="successMission"
         >
           ผู้เข้าร่วมที่สำเร็จ{" "}
           {`(${
@@ -314,7 +377,7 @@ function MissionReport() {
           })`}
         </Radio.Button>
       </Radio.Group>
-      {type === "unsuccess" && (
+      {(type === "unsuccessMission" || type === "unconfirmMission") && (
         <div className="row justify-content-center pt-2">
           <Input
             style={{
@@ -341,7 +404,7 @@ function MissionReport() {
           </Button>
         </div>
       )}
-      {type === "success" && (
+      {type === "successMission" && (
         <div className="row justify-content-center pt-2">
           <Input
             style={{
@@ -376,11 +439,13 @@ function MissionReport() {
           </Button>
         </div>
       )}
-      <Table
+      <NewTable
         className="pt-3"
         columns={columns}
         dataSource={dataMission?.data}
         pagination={false}
+        loading={isLoading}
+        colors={mapTableColor[type]}
       />
       <div className="d-flex justify-content-between pt-3 pb-3">
         <p>รายการทั้งหมด {dataMission?.data.length} รายการ</p>